Moves decorator that skips testing base classes so it can be reused.
